requirements.txt
pandas>=1.5.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.0.0
python-docx>=0.8.11
pdfminer.six>=20221105
beautifulsoup4>=4.11.0
requests>=2.28.0
supabase>=1.0.0
pathlib2>=2.3.7
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
